moved api functions to api dir. Issue #6
